requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==21.4.0
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
-e git+ssh://git@github.com/LLSCskier/UofC2022.git@03cb85a588f3ef9f9e0b6f333c7534d14befbf3f#egg=comp_unet
cycler==0.11.0
dill==0.3.5.1
distlib==0.3.4
filelock==3.7.0
flake8==4.0.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.46.1
h5py==2.10.0
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
Keras==2.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
llvmlite==0.38.1
Markdown==3.3.7
matplotlib==3.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
numba==0.55.2
numpy==1.18.5
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pathlib==1.0.1
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
promise==2.3
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.4.1
sigpy==0.1.23
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
